requirements.txt
python-telegram-bot==21.9
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
apscheduler==3.10.4
pytz==2023.3
